added composer require packages and also changed the migration from integer to biginteger

+++ src/Database/Migrations/2019_08_07_112111_create_dz_campaign_details_table.php
@@ -22,7 +22,7 @@ class CreateDzCampaignDetailsTable extends Migration
         {
             $userTable = LaravelUtility::getUserTable();
 
-            $table->increments('id');
+            $table->bigIncrements('id');
 
             $table->string('name')->comment('Campaign name')->nullable();
             $table->string('description')->comment('Campaign description')->nullable();
@@ -34,8 +34,8 @@ class CreateDzCampaignDetailsTable extends Migration
             $table->boolean('active')->comment('Active flag')->nullable();
             $table->boolean('reusable')->comment('Reusable flag')->nullable();
 
-            $table->unsignedInteger('created_by')->nullable();
-            $table->unsignedInteger('updated_by')->nullable();
+            $table->unsignedBigInteger('created_by')->nullable();
+            $table->unsignedBigInteger('updated_by')->nullable();
 
             $table->softDeletes();
             $table->timestamps();
